refactor(macos): isolate core lifecycle and proxy coordination
